# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beauty Salon Scheduling API",
        "description": "RESTful API for managing beauty salon appointments. Allows listing available time slots, booking appointments, and canceling existing bookings.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/slots": {
            "get": {
                "tags": [
                    "Slots"
                ],
                "summary": "List available appointment slots",
                "description": "Returns all available time slots for a given service and specialist on a specific date. Slots are calculated based on working hours (09:00-18:00) and existing appointments.",
                "operationId": "c062b465d107ee5eb3bd000215d43c4c",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to check availability (YYYY-MM-DD format)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-15"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "ID of the service",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "specialist_id",
                        "in": "query",
                        "description": "ID of the specialist",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of available slots",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "specialist_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-15T09:00:00+00:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-15T09:50:00+00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/book": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Book an appointment",
                "description": "Creates a new appointment for a given specialist, service, and time slot. The slot must be available and within working hours.",
                "operationId": "2d30b393e9d0c60d0904d9caf4e1fb27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "service_id",
                                    "specialist_id",
                                    "start_time"
                                ],
                                "properties": {
                                    "date": {
                                        "description": "Appointment date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-15"
                                    },
                                    "service_id": {
                                        "description": "ID of the service",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "specialist_id": {
                                        "description": "ID of the specialist",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "description": "Start time (HH:MM format)",
                                        "type": "string",
                                        "example": "09:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "specialist_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "service_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "end_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "canceled": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token"
                    },
                    "409": {
                        "description": "Conflict - Slot no longer available"
                    },
                    "422": {
                        "description": "Validation error or outside working hours"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/appointments/{appointment}": {
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Cancel an appointment",
                "description": "Cancels an existing appointment by marking it as canceled. The slot becomes available for future bookings.",
                "operationId": "8502b2f26c87589c313939929663582e",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "description": "Appointment ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment canceled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Canceled"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token"
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your API token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Appointments",
            "description": "Appointment management endpoints"
        },
        {
            "name": "Slots",
            "description": "Available time slots endpoints"
        }
    ]
}